fix: Styling bugs  & conditional rendering
