+++ src/pages/SingleMovie/SIngleMovieDetails.css
@@ -164,12 +164,14 @@
   background-color: #fff;
   font-weight: bold;
   position: absolute;
-  position: absolute;
   /* top: 100%; */
   bottom: 3%;
   right: 45%;
   border-radius: 5px;
   transition: all 0.5s;
+}
+.details-right .showmore-btn#unset-my-ass {
+  position: unset;
 }
 .details-right .showmore-btn.dark-color {
   background-color: #121212;
